# doc/api.yaml
openapi: 3.0.3
info:
  title: "WaSApp"
  description: "Messaging App"
  version: 1.0.0
security:
- bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      title: User
      description: "Represents a user in the system."
      type: object
      properties:
        userId:
          description: "Unique identifier for a user"
          type: integer
          readOnly: true
        name:
          description: "Username"
          type: string
          pattern: "^.*$"
          minLength: 3
          maxLength: 16
        profilePicture:
          description: "Profile picture of a user (Base64 encoded)"
          type: string
          format: byte
          pattern: "^.*$"
          minLength: 1
          maxLength: 100000
      required: [ "userId", "name" ]
...
    Error:
      title: Error
      description: "Error response object."
      type: object
      properties:
        error:
          type: string
          description: "Human-readable error message"
          pattern: "^.*$"
          minLength: 1
          maxLength: 200
      required: [ "error" ]
  parameters:
    userId:
      schema:
        type: integer
        readOnly: true
      name: id
      in: path
      required: true
      description: "User Identifier"
...
paths:
  /session:
    post:
      tags: [ "login" ]
      summary: "Logs in the user"
      description: "If the user does not exist, it will be created, and an identifier is returned. If the user exists, the user identifier is returned."
      operationId: doLogin
      requestBody:
        description: "User details required for login."
        required: true
        content:
          application/json:
            schema:
              type: object
              description: "Login input schema."
              properties:
                name:
                  type: string
                  description: "Name of the user logging in."
                  pattern: "^.*$"
                  minLength: 3
                  maxLength: 16
              required: [ "name" ]
      responses:
        '201':
          description: "User log-in action successful"
          content:
            application/json:
              schema:
                type: object
                description: "Response containing the user's identifier."
                properties:
                  identifier:
                    type: string
                    description: "User's unique login identifier."
                    pattern: "^.*$"
                    minLength: 1
                    maxLength: 50
                required: [ "identifier" ]
        '400':
          description: "Invalid input (e.g., name too short)"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /users:
    get:
      tags: [ "users" ]
      summary: "List all users or search users by name"
      description: "Retrieves a list of users. If a `name` query parameter is provided, returns only users whose names match or contain the given string."
      operationId: listUsers
      parameters:
      - name: name
        in: query
        description: "Filter users by name (partial matches allowed)."
        required: false
        schema:
          type: string
          pattern: "^.*$"
          minLength: 1
          maxLength: 50
      responses:
        "200":
          description: "List of users (possibly filtered by name)."
          content:
            application/json:
              schema:
                type: array
                description: "An array of user objects."
                items:
                  $ref: "#/components/schemas/User"
                minItems: 0
                maxItems: 10000
        "400":
          description: "Bad Request - Invalid query parameter"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /users/{id}:
    parameters:
    - $ref: "#/components/parameters/userId"
    put:
      tags: [ "users" ]
      summary: "Change username by id"
      description: "Change your own username to a new value."
      operationId: setMyUserName
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: "Schema for updating the username."
              properties:
                newName:
                  type: string
                  description: "New name to be assigned."
                  pattern: "^.*$"
                  minLength: 3
                  maxLength: 16
              required: [ "newName" ]
      responses:
        '200':
          description: "User name updated successfully"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        '400':
          description: "Bad Request - Invalid input"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '404':
          description: "User not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '409':
          description: "Conflict - New username already in use"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...